hardware/requirements.txt
# pip packages
requests==2.31.0
numpy==1.24.4
opencv-python==4.8.1.78
face-recognition==1.3.0
pyserial==3.5
adafruit-circuitpython-fingerprint==2.2.10
mfrc522==0.0.7
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus2==0.4.3

# NOTE: picamera2 is NOT installed via pip on Raspberry Pi.
# Install it via apt (it's pre-installed on Raspberry Pi OS Bookworm):
#   sudo apt install -y python3-picamera2 python3-libcamera
